chore: I extracted the code and turned it into a component
